Move original computed function to watcher and simplify
