+++ tests/test_store.py
@@ -2,7 +2,8 @@ from unittest.mock import Mock
 
 import pytest
 
-from observ import computed, mutation, Store, watch
+from observ import watch
+from observ.store import computed, mutation, Store
 
 
 class CustomStore(Store):
@@ -54,30 +55,13 @@ def test_store_undo_redo():
 def test_store_computed_methods():
     store = CustomStore(state={"count": 0})
 
-    @computed
-    def double():
-        return store.state["count"] * 2
-
-    def triple_count():
-        return store.state["count"] * 3
-
-    triple = computed(triple_count)
-
-    quadruple = computed(lambda: store.state["count"] * 4)
-
     assert store.state["count"] == 0
     assert store.double == 0
-    assert double() == 0
-    assert triple() == 0
-    assert quadruple() == 0
 
     store.bump_count()
 
     assert store.state["count"] == 1
     assert store.double == 2
-    assert double() == 2
-    assert triple() == 3
-    assert quadruple() == 4
 
 
 @pytest.mark.xfail(reason="deepcopy is used to restore state")
